Create View Factor complete

+++ src/main/java/com/darash/salemaven/beans/FactorDetailController.java
@@ -140,7 +140,7 @@ public class FactorDetailController implements Serializable {
             return key;
         }
 
-        String getStringKey(java.lang.Long value) {
+        String getStringKey(java.lang.Integer value) {
             StringBuilder sb = new StringBuilder();
             sb.append(value);
             return sb.toString();
